update deprecated code (linear_assignment and np's type)

+++ application_util/preprocessing.py
@@ -37,7 +37,7 @@ def non_max_suppression(boxes, max_bbox_overlap, scores=None):
     if len(boxes) == 0:
         return []
 
-    boxes = boxes.astype(np.float)
+    boxes = boxes.astype(np.float32)
     pick = []
 
     x1 = boxes[:, 0]
